Add `aws-neuron-runtime-base` to list of yum packages in release notes
The release notes contain instructions for updating a DLAMI machine to
the latest version of the neuron software.  However, on Amazon Linux,
the instructions do not mention the `aws-neuron-runtime-base` package.
I found that running the `yum` commands as specified in the release
notes would upgrade the other neuron rpms, but not the `runtime-base`
one, even though there was a new version available.

I did not test this on Ubuntu, am not sure whether `apt` is smart
enough to recognize the dependency on `aws-neuron-runtime-base` and
update it as part of the execution of the commands mentioned in the
release notes.