Neuron SDK Release - 1.16.1 - November 5 2021
Neuron 1.16.1 is a patch release. This release fixes a bug in Neuron Runtime that would have prevented users from launching a container that doesn’t use all of the Neuron Devices in the instance. If you are using Neuron within a container, please update to this new release by updating to latest Neuron ML framework package, Neuron Tools, and/or TensorFlow Neuron Model Server.

Co-authored-by: micwade (AWS) <57163979+micwade-aws@users.noreply.github.com>